Remove unnecessary trait, manual require, move to dedicated helpers dir
