+++ src/tests/bootstrap.php
@@ -32,7 +32,6 @@ function qit_tests_reset_config_dir() {
 
 require_once __DIR__ . '/../vendor/autoload.php';
 require_once __DIR__ . '/../src/helpers.php';
-require_once __DIR__ . '/ZipBuilderHelper.php';
 
 qit_tests_reset_config_dir();
